Fix comparisons of integers with different sizes and signedness
Use size_t for loop indices compared against container sizes,
add static_casts where signed/unsigned comparisons are needed,
and fix related integer type warnings.

- Adjusting loop variable types to match container size types
- Adding explicit static_cast<> for necessary signed/unsigned comparisons
- Adding pragma to suppress intentional mixed-sign comparisons in
  Expression_evaluate_binary_comparison.cpp

+++ shared/rocroller/client/include/client/DataParallelGEMMSolution.hpp
@@ -355,11 +355,11 @@ namespace rocRoller
 
                     AssertFatal(solutionParams.macM * solutionParams.macK
                                         * DataTypeInfo::Get(typeA).elementBytes
-                                    > wave_m * wave_k,
+                                    > static_cast<unsigned int>(wave_m * wave_k),
                                 "Not enough elements (A).");
                     AssertFatal(solutionParams.macN * solutionParams.macK
                                         * DataTypeInfo::Get(typeA).elementBytes
-                                    > wave_n * wave_k,
+                                    > static_cast<unsigned int>(wave_n * wave_k),
                                 "Not enough elements (B).");
 
                     auto const arch = GPUArchitectureLibrary::getInstance()->GetArch(
