+++ shared/rocroller/lib/source/ExpressionTransformations/Simplify.cpp
@@ -154,7 +154,7 @@ namespace rocRoller
                     //
                     // Literal 0 can only accept Int32/UInt32/Int64/UInt64/Half/Float/Double
                     //
-                    if((elementSize == 4u or elementSize == 8u) and (rhs >= elementSize * 8u))
+                    if((elementSize == 4u or elementSize == 8u) and (static_cast<size_t>(rhs) >= elementSize * 8u))
                     {
                         return literal(0, resultVarType);
                     }
@@ -636,7 +636,7 @@ namespace rocRoller
                 std::vector<ExpressionPtr> overlapOperands;
                 uint32_t                   firstOperandStartBit = 0;
 
-                for(int i = 0; i < expr.operands.size(); ++i)
+                for(size_t i = 0; i < expr.operands.size(); ++i)
                 {
                     operandStartBit = operandEndBit;
                     operandEndBit += resultVariableType(expr.operands[i]).getElementSize() * 8;
@@ -846,7 +846,7 @@ namespace rocRoller
 
                 // Extracting the entire arg with no offset
                 auto argVarType = resultVariableType(cpy.arg);
-                if(cpy.offset == 0 && cpy.width == argVarType.getElementSize() * 8
+                if(cpy.offset == 0 && static_cast<size_t>(cpy.width) == argVarType.getElementSize() * 8
                    && argVarType.getElementSize()
                           == DataTypeInfo::Get(cpy.outputDataType).elementBytes)
                     return call(reinterpret(cpy.outputDataType, cpy.arg));
